"Enable 'cabe' plugin and set JVM options in build.gradle.kts
Enabled the previously commented out 'com.dua3.cabe' plugin in build.gradle.kts file again. By setting the system property for prism.order in the compileJava task to 2j2d", we circumvent the problem of the build failing due to not being able to load the dynlib for the renderer at build time.